move parts into src/protoc
follows our convention
