+++ lib/protoc.dart
@@ -15,15 +15,15 @@ import 'src/dart_options.pb.dart';
 import 'const_generator.dart' show writeJsonConst;
 import 'indenting_writer.dart' show IndentingWriter;
 
-part 'base_type.dart';
-part 'client_generator.dart';
-part 'code_generator.dart';
-part 'enum_generator.dart';
-part 'extension_generator.dart';
-part 'file_generator.dart';
-part 'linker.dart';
-part 'message_generator.dart';
-part 'options.dart';
-part 'output_config.dart';
-part 'protobuf_field.dart';
-part 'service_generator.dart';
+part 'src/protoc/base_type.dart';
+part 'src/protoc/client_generator.dart';
+part 'src/protoc/code_generator.dart';
+part 'src/protoc/enum_generator.dart';
+part 'src/protoc/extension_generator.dart';
+part 'src/protoc/file_generator.dart';
+part 'src/protoc/linker.dart';
+part 'src/protoc/message_generator.dart';
+part 'src/protoc/options.dart';
+part 'src/protoc/output_config.dart';
+part 'src/protoc/protobuf_field.dart';
+part 'src/protoc/service_generator.dart';
