+++ lib/src/protoc/client_generator.dart
