+++ lib/src/protoc/enum_generator.dart
