+++ lib/src/protoc/extension_generator.dart
