+++ lib/src/protoc/file_generator.dart
