+++ lib/src/protoc/linker.dart
