+++ lib/src/protoc/message_generator.dart
